Moving classes to logically more meaningful package.
