+++ components/client/src/test/unit/java/com/hotels/styx/client/netty/connectionpool/FlowControllingHttpContentProducerTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2013-2017 Expedia Inc.
+ * Copyright (C) 2013-2018 Expedia Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hotels.styx.client.netty;
+package com.hotels.styx.client.netty.connectionpool;
 
 import com.hotels.styx.api.client.Origin;
+import com.hotels.styx.client.netty.ConsumerDisconnectedException;
+import com.hotels.styx.client.netty.connectionpool.FlowControllingHttpContentProducer;
 import com.hotels.styx.support.matchers.LoggingTestSupport;
 import com.hotels.styx.api.netty.exceptions.ResponseTimeoutException;
 import com.hotels.styx.api.netty.exceptions.TransportLostException;
@@ -34,7 +36,7 @@ import static ch.qos.logback.classic.Level.WARN;
 import static com.google.common.base.Charsets.UTF_8;
 import static com.hotels.styx.api.client.Origin.newOriginBuilder;
 import static com.hotels.styx.support.matchers.LoggingEventMatcher.loggingEvent;
-import static com.hotels.styx.client.netty.FlowControllingHttpContentProducer.ProducerState.*;
+import static com.hotels.styx.client.netty.connectionpool.FlowControllingHttpContentProducer.ProducerState.*;
 import static io.netty.buffer.Unpooled.copiedBuffer;
 import static java.util.Collections.emptyList;
 import static org.hamcrest.MatcherAssert.assertThat;
